pyproject.toml
[project]
name = "rinha-de-backend-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "django>=5.2.4",
    "django-ninja>=1.4.3",
    "httpx>=0.28.1",
    "msgpack>=1.1.1",
    "psycopg>=3.2.9",
    "redis>=6.4.0",
    "uvicorn>=0.35.0",
]
